+++ CHANGELOG.md
@@ -10,7 +10,11 @@ app.
 
 - Added an option to use system fonts instead of bundled fonts
   ([#1799](https://github.com/birchill/10ten-ja-reader/issues/1799)).
+- (Safari) Fixed an issue with the popup stalling when looking up words using
+  the puck on
+  ([#1802](https://github.com/birchill/10ten-ja-reader/issues/1802)).
 - Fixed kanji incorrectly encoded as radicals not being displayed in kanji tab
+  thanks to [@Enellis](https://github.com/Enellis)
   ([#1205](https://github.com/birchill/10ten-ja-reader/issues/1205#issuecomment-1783641017)).
 
 ## [1.19.0] - 2024-05-30